# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIS Kernel Control Daemon (sisctl)",
    "description": "REST API for managing SIS kernel QEMU instances",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8871",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/v1/autonomy/interval": {
      "post": {
        "tags": [
          "autonomy"
        ],
        "summary": "Set autonomy interval",
        "operationId": "autonomySetInterval",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "interval_ms": {
                    "type": "integer"
                  }
                },
                "required": [
                  "interval_ms"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interval updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutonomyStatus"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/autonomy/conf-threshold": {
      "post": {
        "tags": [
          "autonomy"
        ],
        "summary": "Set confidence threshold",
        "operationId": "autonomySetThreshold",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "threshold": {
                    "type": "integer"
                  }
                },
                "required": [
                  "threshold"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Threshold updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutonomyStatus"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/autonomy/audit": {
      "get": {
        "tags": [
          "autonomy"
        ],
        "summary": "Get autonomy audit log",
        "operationId": "autonomyAudit",
        "parameters": [
          {
            "name": "last",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Audit entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutonomyDecision"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/autonomy/explain": {
      "get": {
        "tags": [
          "autonomy"
        ],
        "summary": "Explain a decision",
        "operationId": "autonomyExplain",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Decision explanation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/autonomy/preview": {
      "post": {
        "tags": [
          "autonomy"
        ],
        "summary": "Preview next decisions",
        "operationId": "autonomyPreview",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preview results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/autonomy/whatif": {
      "post": {
        "tags": [
          "autonomy"
        ],
        "summary": "What-if scenario analysis",
        "operationId": "autonomyWhatIf",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "What-if results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatIfResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutonomyStatus": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "safe_mode": {
            "type": "boolean"
          },
          "learning_frozen": {
            "type": "boolean"
          },
          "interval_ms": {
            "type": "integer"
          },
          "threshold": {
            "type": "integer"
          },
          "total_decisions": {
            "type": "integer"
          },
          "accepted": {
            "type": "integer"
          },
          "deferred": {
            "type": "integer"
          },
          "watchdog_low_reward": {
            "type": "integer"
          },
          "watchdog_high_td_error": {
            "type": "integer"
          }
        },
        "required": [
          "enabled",
          "interval_ms",
          "threshold",
          "total_decisions"
        ]
      },
      "AutonomyDecision": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "timestamp_us": {
            "type": "integer"
          },
          "action": {
            "type": "string"
          },
          "confidence": {
            "type": "integer"
          },
          "reward": {
            "type": "number",
            "format": "float"
          },
          "executed": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "timestamp_us",
          "action",
          "confidence"
        ]
      },
      "AttentionWeight": {
        "type": "object",
        "properties": {
          "feature": {
            "type": "string"
          },
          "weight": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "feature",
          "weight"
        ]
      },
      "ExplainResponse": {
        "type": "object",
        "properties": {
          "decision_id": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "reasoning": {
            "type": "string"
          },
          "attention": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttentionWeight"
            }
          },
          "context": {
            "type": "object"
          }
        },
        "required": [
          "decision_id",
          "explanation"
        ]
      },
      "PreviewRequest": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "count"
        ]
      },
      "PreviewResponse": {
        "type": "object",
        "properties": {
          "decisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AutonomyDecision"
            }
          },
          "would_execute": {
            "type": "boolean"
          },
          "confidence": {
            "type": "integer"
          }
        },
        "required": [
          "decisions"
        ]
      },
      "WhatIfRequest": {
        "type": "object",
        "properties": {
          "overrides": {
            "type": "object"
          }
        },
        "required": [
          "overrides"
        ]
      },
      "WhatIfResponse": {
        "type": "object",
        "properties": {
          "scenario": {
            "$ref": "#/components/schemas/PreviewResponse"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "scenario"
        ]
      },
...
    }
  },
...
}